pc-gui-agent/requirements.txt
ollama>=0.1.0
playwright>=1.40.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
flet>=0.21.0

# MCP Python SDK 依赖
# 方法1：使用本地 python-sdk（推荐，如果已克隆）
# pip install -e ./python-sdk
# 
# 方法2：从 PyPI 安装官方包
# pip install mcp
# 
# 注意：代码会自动尝试从本地 python-sdk 导入，如果失败则从已安装的包导入
# 详细安装说明请参考: docs/MCP_SETUP.md

